trunk_endpoint_model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
open3d>=0.18.0
tqdm>=4.65.0
wandb>=0.15.0
scipy>=1.10.0
Pillow>=9.0.0
opencv-python>=4.7.0

